Refactor to support modular effect components

+++ src/luxaudio/analyzers/analyzer.go
@@ -1,5 +1,5 @@
 package analyzers
 
 type Analyzer interface {
-	Analyze([]float64) []byte
+	Analyze([]float64) []float64
 }
